@@ -1,8 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
+//import VueRouter from 'vue-router';
+import ElementPlus from 'element-plus'
+import 'element-plus/theme-chalk/index.css'
+import router from './router/index.js'
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).use(ElementPlus).use(router).mount('#app')
